refactor(test-intelligence): extract domain modules for maintainability
Refactored test-intelligence.js (847 LOC) into focused domain modules:

- validators.js (331 LOC): Main validation functions (phases 1.1-1.4)
- extractors.js (298 LOC): Selector, navigation, component extraction
- file-utils.js (227 LOC): File and branch-aware utilities
- error-analysis.js (168 LOC): Error classification and fix generation
- index.js (88 LOC): Main orchestrator with re-exports

Original file reduced to thin wrapper (52 LOC) maintaining backward compatibility.
Part of SD-LEO-REFACTOR-LARGE-FILES-003 orchestrator.
